MNT: move load_configuration to top level
